Galeri: explicitly cast argument to abs to double since abs(long long) missing. Fix for Linux-PGI-11.1-SERIAL_OPT_DEV1 Windows_NT-MSVC9-SERIAL_OPT_DEV
Build/Test Cases Summary
Enabled Packages: Epetra, Teuchos
Disabled Packages: SEACAS,Sundance,STK
Enabled all Forward Packages
0) MPI_DEBUG => FAILED: passed=1729,notpassed=64 => Not ready to push! (88.31 min)
1) SERIAL_RELEASE => passed: passed=1700,notpassed=0 (56.03 min)
WARNING: Forced the push!

+++ packages/galeri/src/CrsMatrices/Galeri_Fielder.h
@@ -73,7 +73,7 @@ Epetra_CrsMatrix* Fielder(const Epetra_Map* Map)
     for (int_type j = 0 ; j < NumGlobalElements ; ++j) 
     {
       Indices[j] = j;
-      Values[j]  = (double)abs(iGlobal - j);
+      Values[j]  = (double)abs((double) (iGlobal - j));
     }
 
     Matrix->InsertGlobalValues(MyGlobalElements[i], NumEntries,
