+++ packages/galeri/src/CrsMatrices/Galeri_KMS.h
@@ -72,7 +72,7 @@ Epetra_CrsMatrix* KMS(const Epetra_Map* Map, const double value)
     for (int_type j = 0 ; j < NumGlobalElements ; ++j) 
     {
       Indices[j] = j;
-      Values[j] = pow(value, (double)(abs(iGlobal - j)));
+      Values[j] = pow(value, (double)(abs((double) (iGlobal - j))));
     }
 
     Matrix->InsertGlobalValues(MyGlobalElements[i], NumEntries,
